pyproject.toml
[project]
name = "crust-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.9.0",
    "marimo>=0.12.2",
    "nltk>=3.9.1",
    "selenium>=4.30.0",
    "termcolor>=3.0.1",
]
